fix(explorer): remove flickering from app-data details
