+++ apps/explorer/src/explorer/pages/AppData/DecodePage.tsx
@@ -4,7 +4,7 @@ import Form, { FormValidation } from '@rjsf/core'
 
 import { decodeAppDataSchema, FormProps, handleErrors, transformErrors } from './config'
 
-import DecodeAppData from '../../../components/AppData/DecodeAppData'
+import { DecodeAppData } from '../../../components/AppData/DecodeAppData'
 
 import { TabData } from './index'
 
@@ -13,15 +13,13 @@ type DecodeProps = {
   setTabData: React.Dispatch<React.SetStateAction<TabData>>
 }
 
-// TODO: Break down this large function into smaller functions
-// eslint-disable-next-line max-lines-per-function
 const DecodePage: React.FC<DecodeProps> = ({ tabData, setTabData }) => {
   const { decode } = tabData
   const [formData, setFormdata] = useState<FormProps>(decode.formData)
   const [isSubmitted, setIsSubmitted] = useState<boolean>(decode.options.isSubmitted ?? false)
   const [disabled, setDisabled] = useState<boolean>(decode.options.disabled ?? true)
   const [invalidFormDataAttempted, setInvalidFormDataAttempted] = useState<boolean>(
-    decode.options.invalidFormDataAttempted ?? false
+    decode.options.invalidFormDataAttempted ?? false,
   )
 
   useEffect(() => {
